requirements.txt
confluent-kafka
fastavro
avro-python